added download for cover letter
