+++ src/index.tsx
@@ -11,7 +11,7 @@ import { HomePage } from "./pages/HomePage";
 import { SelectPersonaPage } from "./pages/performance-review/SelectPersonaPage";
 import { PerformanceReviewPage } from "./pages/performance-review/PerformanceReviewPage";
 import { ResumeSelectionPage } from "./pages/resumes/ResumeSelectionPage";
-import { CoverLetterPage } from "./pages/resumes/CoverLetterPage";
+import { CoverLetterPage } from "./pages/resumes/cover/CoverLetterPage";
 import { ReferralPage } from "./pages/resumes/referral/ReferralPage";
 import { ResumePage } from "./pages/resumes/cv/ResumePage";
 import { PrivacyPolicyPage } from "./pages/PrivacyPolicyPage";
